pyproject.toml
[tool.poetry]
name = "cpl-explore"
version = "0.1.0"
description = ""
authors = ["karanrajagopalan <kir@equinor.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cognite-sdk = {version = ">=7", extras = ["pandas"]}
streamlit = "^1.41.1"
streamlit-folium = "^0.23.2"
shapely = "^2.0.6"
folium = "^0.19.2"
dash = "^2.18.2"
dash-leaflet = "^1.0.15"
dash-extensions = "^1.0.19"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
arrow = "^1.3.0"
pydantic-settings = "^2.7.1"
pre-commit = "^4.1.0"
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"
mypy = "^1.14.1"
openpyxl = "^3.1.5"
pyodbc = "^5.2.0"
dotenv = "^0.9.9"
sqlalchemy = "^2.0.43"
geopandas = "^1.1.1"
geojson = "^3.2.0"
geobuf = "^1.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
